Changes in endpoints and classes: - (/download-encrypted-text): New endpoint created, download the sent content - (/encrypt): Modify the if statement to remove the created file - FilesUtil: Delete "createFile" function with byte array like argument because it's unused

+++ src/main/java/io/ivndot/routes/Encrypt.java
@@ -96,8 +96,9 @@ public class Encrypt extends HttpServlet {
 		}
 
 		// delete file
-		if (fileToEncrypt.exists())
-			fileToEncrypt.delete();
+		if (fileToEncrypt != null)
+			if (fileToEncrypt.exists())
+				fileToEncrypt.delete();
 
 		// send response
 		ResponseUtil.sendJSONResponse(resp, encryptBean, "POST");
